#544: split bountycontrolpanel into separate files
